update: table laundry kolom nomer

+++ database/migrations/2023_02_06_073846_laundy_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->increments('id');
             $table->string('nama');
             $table->string('alamat');
-            $table->integer('nomerhp');
+            $table->BigInteger('nomerhp');
             $table->integer('berat');
             $table->integer('jenis_paket');
             $table->timestamps();
